backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
psycopg2-binary
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
apscheduler
